Add an environment flag for LILAC_AUTH_ENABLED and turn off features. (#452)
Features turned off with readonly:
- Create a dataset
- Delete a dataset or concept
- Computing a signal
- Computing an embedding
- Changing dataset settings
- Monitoring tasks (there are no tasks in readonly)

Features *not yet turned off* (will do this once I figure out auth in
HF):
- Concept editing (thumbs up / down)
- Creating a concept

Readonly demo: https://huggingface.co/spaces/lilacai/nikhil_staging

+++ src/router_concept.py
@@ -7,6 +7,7 @@ from fastapi import APIRouter, HTTPException
 from openai_function_call import OpenAISchema
 from pydantic import BaseModel, Field
 
+from .auth import get_user_access
 from .concepts.concept import (
   DRAFT_MAIN,
   Concept,
@@ -81,6 +82,9 @@ def edit_concept(namespace: str, concept_name: str, change: ConceptUpdate) -> Co
 @router.delete('/{namespace}/{concept_name}')
 def delete_concept(namespace: str, concept_name: str) -> None:
   """Deletes the concept from the database."""
+  if not get_user_access().concept.delete_any_concept:
+    raise HTTPException(401, 'User does not have access to delete this concept.')
+
   DISK_CONCEPT_DB.remove(namespace, concept_name)
   # Delete concept models from all datasets that are using this concept.
   DISK_CONCEPT_MODEL_DB.remove_all(namespace, concept_name)
