+++ src/router_dataset.py
@@ -2,10 +2,11 @@
 from typing import Optional, Sequence, Union, cast
 from urllib.parse import unquote
 
-from fastapi import APIRouter, Response
+from fastapi import APIRouter, HTTPException, Response
 from fastapi.responses import ORJSONResponse
 from pydantic import BaseModel, validator
 
+from .auth import get_user_access
 from .config import data_path
 from .data.dataset import BinaryOp
 from .data.dataset import Column as DBColumn
@@ -82,6 +83,9 @@ class ComputeSignalOptions(BaseModel):
 @router.delete('/{namespace}/{dataset_name}')
 def delete_dataset(namespace: str, dataset_name: str) -> None:
   """Delete the dataset."""
+  if not get_user_access().dataset.delete_dataset:
+    raise HTTPException(401, 'User does not have access to delete this dataset.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.delete()
   remove_dataset_from_cache(namespace, dataset_name)
@@ -96,6 +100,8 @@ class ComputeSignalResponse(BaseModel):
 def compute_signal(namespace: str, dataset_name: str,
                    options: ComputeSignalOptions) -> ComputeSignalResponse:
   """Compute a signal for a dataset."""
+  if not get_user_access().dataset.compute_signals:
+    raise HTTPException(401, 'User does not have access to compute signals over this dataset.')
 
   def _task_compute_signal(namespace: str, dataset_name: str, options_dict: dict,
                            task_id: TaskId) -> None:
@@ -130,6 +136,9 @@ class DeleteSignalResponse(BaseModel):
 def delete_signal(namespace: str, dataset_name: str,
                   options: DeleteSignalOptions) -> DeleteSignalResponse:
   """Delete a signal from a dataset."""
+  if not get_user_access().dataset.delete_signals:
+    raise HTTPException(401, 'User does not have access to delete this signal.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.delete_signal(options.signal_path)
   return DeleteSignalResponse(completed=True)
@@ -293,6 +302,9 @@ def get_settings(namespace: str, dataset_name: str) -> DatasetSettings:
 @router.post('/{namespace}/{dataset_name}/settings', response_model_exclude_none=True)
 def update_settings(namespace: str, dataset_name: str, settings: DatasetSettings) -> None:
   """Get the media for the dataset."""
+  if not get_user_access().dataset.compute_signals:
+    raise HTTPException(401, 'User does not have access to update the settings of this dataset.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.update_settings(settings)
   return None
